fix/finish audio service deprecation
